LayoutTests/fast/regions/webkit-flow-parsing-expected.txt

Test parsing of the CSS webkit-flow-into property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test("-webkit-flow-into: auto") is "auto"
PASS test("-webkit-flow-into: first-flow") is "first-flow"
PASS test("-webkit-flow-into: 'first flow'") is ""
PASS test("-webkit-flow-into: ;") is ""
PASS test("-webkit-flow-into: 1") is ""
PASS test("-webkit-flow-into: 1.2") is ""
PASS test("-webkit-flow-into: -1") is ""
PASS test("-webkit-flow-into: 12px") is ""
PASS testComputedStyle("auto") is "auto"
PASS testComputedStyle("") is "auto"
PASS testComputedStyle("'first-flow'") is "auto"
PASS testComputedStyle("first-flow") is "first-flow"
PASS testComputedStyle("12px") is "auto"
PASS testNotInherited("auto", "auto") is "auto"
PASS testNotInherited("auto", "child-flow") is "child-flow"
PASS testNotInherited("parent-flow", "auto") is "auto"
PASS testNotInherited("parent-flow", "child-flow") is "child-flow"
PASS successfullyParsed is true

TEST COMPLETE
